+++ frontend/src/forms/Setup.vue
@@ -9,7 +9,7 @@
                 <div class="row">
                     <div class="col-6">
                         <div class="form-group">
-                            <label>{{ $t('setup.language') }}</label>
+                            <label class="text-capitalize">{{ $t('setup.language') }}</label>
                             <select @change="changeLanguages" v-model="setup.language" id="language" class="form-control">
                                 <option value="en">English</option>
                                 <option value="es">Spanish</option>
@@ -19,7 +19,7 @@
                             </select>
                         </div>
                         <div class="form-group">
-                            <label>{{ $t('setup.connection') }}</label>
+                            <label class="text-capitalize">{{ $t('setup.connection') }}</label>
                             <select @change="canSubmit" v-model="setup.db_connection" id="db_connection" class="form-control">
                                 <option value="sqlite">SQLite</option>
                                 <option value="postgres">Postgres</option>
@@ -27,23 +27,23 @@
                             </select>
                         </div>
                         <div v-if="setup.db_connection !== 'sqlite'" class="form-group">
-                            <label>{{ $t('setup.host') }}</label>
+                            <label class="text-capitalize">{{ $t('setup.host') }}</label>
                             <input @keyup="canSubmit" v-model="setup.db_host" id="db_host" type="text" class="form-control" placeholder="localhost">
                         </div>
                         <div v-if="setup.db_connection !== 'sqlite'" class="form-group">
-                            <label>{{ $t('port') }}</label>
+                            <label class="text-capitalize">{{ $t('port') }}</label>
                             <input @keyup="canSubmit" v-model="setup.db_port" id="db_port" type="text" class="form-control" placeholder="localhost">
                         </div>
                         <div v-if="setup.db_connection !== 'sqlite'" class="form-group">
-                            <label>{{ $t('username') }}</label>
+                            <label class="text-capitalize">{{ $t('username') }}</label>
                             <input @keyup="canSubmit" v-model="setup.db_user" id="db_user" type="text" class="form-control" placeholder="root">
                         </div>
                         <div v-if="setup.db_connection !== 'sqlite'" class="form-group">
-                            <label for="db_password">{{ $t('password') }}</label>
+                            <label for="db_password" class="text-capitalize">{{ $t('password') }}</label>
                             <input @keyup="canSubmit" v-model="setup.db_password" id="db_password" type="password" class="form-control" placeholder="password123">
                         </div>
                         <div v-if="setup.db_connection !== 'sqlite'" class="form-group">
-                            <label for="db_database">{{ $t('database') }}</label>
+                            <label for="db_database" class="text-capitalize">{{ $t('database') }}</label>
                             <input @keyup="canSubmit" v-model="setup.db_database" id="db_database" type="text" class="form-control" placeholder="Database name">
                         </div>
 
@@ -52,43 +52,43 @@
                     <div class="col-6">
 
                         <div class="form-group">
-                            <label>{{ $t('setup.project_name') }}</label>
+                            <label class="text-capitalize">{{ $t('setup.project_name') }}</label>
                             <input @keyup="canSubmit" v-model="setup.project" id="project" type="text" class="form-control" placeholder="Great Uptime" required>
                         </div>
 
                         <div class="form-group">
-                            <label>{{ $t('setup.project_description') }}</label>
+                            <label class="text-capitalize">{{ $t('setup.project_description') }}</label>
                             <input @keyup="canSubmit" v-model="setup.description" id="description" type="text" class="form-control" placeholder="Great Uptime">
                         </div>
 
                         <div class="form-group">
-                            <label for="domain">{{ $t('setup.domain') }}</label>
+                            <label class="text-capitalize" for="domain">{{ $t('setup.domain') }}</label>
                             <input @keyup="canSubmit" v-model="setup.domain" type="text" class="form-control" id="domain" required>
                         </div>
 
                         <div class="form-group">
-                            <label>{{ $t('setup.username') }}</label>
+                            <label class="text-capitalize">{{ $t('setup.username') }}</label>
                             <input @keyup="canSubmit" v-model="setup.username" id="username" type="text" class="form-control" placeholder="admin" required>
                         </div>
 
                         <div class="form-group">
-                            <label>{{ $t('setup.username') }}</label>
+                            <label class="text-capitalize">{{ $t('setup.username') }}</label>
                             <input @keyup="canSubmit" v-model="setup.password" id="password" type="password" class="form-control" placeholder="password" required>
                         </div>
 
                         <div class="form-group">
-                            <label>{{ $t('setup.password_confirm') }}</label>
+                            <label class="text-capitalize">{{ $t('setup.password_confirm') }}</label>
                             <input @keyup="canSubmit" v-model="setup.confirm_password" id="password_confirm" type="password" class="form-control" placeholder="password" required>
                         </div>
 
                         <div class="form-group">
                             <div class="row">
                                 <div class="col-8">
-                                    <label>{{ $t('email') }}</label>
+                                    <label class="text-capitalize">{{ $t('email') }}</label>
                                     <input @keyup="canSubmit" v-model="setup.email" id="email" type="text" class="form-control" placeholder="myemail@domain.com">
                                 </div>
                                 <div class="col-4">
-                                    <label class="d-none d-sm-block">{{ $t('setup.newsletter') }}</label>
+                                    <label class="d-none d-sm-block text-capitalize text-capitalize">{{ $t('setup.newsletter') }}</label>
                                     <span @click="setup.newsletter = !!setup.newsletter" class="switch">
                                       <input v-model="setup.newsletter" type="checkbox" name="using_cdn" class="switch" id="send_newsletter" :checked="setup.newsletter">
                                       <label for="send_newsletter"></label>