+++ utils/log.go
@@ -146,6 +146,9 @@ func createLog(dir string) error {
 
 // InitLogs will create the '/logs' directory and creates a file '/logs/statup.log' for application logging
 func InitLogs() error {
+	if Params.GetBool("DISABLE_LOGS") {
+		return nil
+	}
 	if err := createLog(Directory); err != nil {
 		return err
 	}
@@ -164,8 +167,6 @@ func InitLogs() error {
 		DisableColors: false,
 	})
 	checkVerboseMode()
-
-	LastLines = make([]*logRow, 0)
 	return nil
 }